+++ optionlab/engine.py
@@ -2,11 +2,9 @@ from __future__ import division
 from __future__ import print_function
 
 import datetime as dt
-from typing import Any
 
-from numpy import array, zeros
+from numpy import zeros
 
-# from numpy import ndarray
 
 from optionlab.black_scholes import get_bs_info, get_implied_vol
 from optionlab.models import (
@@ -17,7 +15,6 @@ from optionlab.models import (
     ClosedPosition,
     Outputs,
     BlackScholesModelInputs,
-    LaplaceInputs,
     ArrayInputs,
     OptionType,
     EngineData,
@@ -26,13 +23,10 @@ from optionlab.support import (
     get_pl_profile,
     get_pl_profile_stock,
     get_pl_profile_bs,
-    get_profit_range,
     create_price_seq,
     get_pop,
 )
 from optionlab.utils import get_nonbusiness_days
-
-# from optionlab.utils import get_pl, pl_to_csv
 
 
 def run_strategy(inputs_data: Inputs | dict) -> Outputs:
@@ -68,11 +62,11 @@ def run_strategy(inputs_data: Inputs | dict) -> Outputs:
 def _init_inputs(inputs: Inputs) -> EngineData:
     data = EngineData(
         stock_price_array=create_price_seq(inputs.min_stock, inputs.max_stock),
-        terminal_stock_prices=inputs.array if inputs.array is not None else array([]),
+        terminal_stock_prices=inputs.array if inputs.model == "array" else None,
         inputs=inputs,
     )
 
-    data._days_in_year = (
+    data.days_in_year = (
         inputs.business_days_in_year if inputs.discard_nonbusiness_days else 365
     )
 
@@ -98,11 +92,11 @@ def _init_inputs(inputs: Inputs) -> EngineData:
             data.premium.append(strategy.premium)
             data.n.append(strategy.n)
             data.action.append(strategy.action)
-            data._previous_position.append(strategy.prev_pos or 0.0)
+            data.previous_position.append(strategy.prev_pos or 0.0)
 
             if not strategy.expiration:
-                data._days_to_maturity.append(data.days_to_target)
-                data._use_bs.append(False)
+                data.days_to_maturity.append(data.days_to_target)
+                data.use_bs.append(False)
             elif isinstance(strategy.expiration, dt.date) and inputs.start_date:
                 if inputs.discard_nonbusiness_days:
                     n_discarded_days = get_nonbusiness_days(
@@ -111,18 +105,18 @@ def _init_inputs(inputs: Inputs) -> EngineData:
                 else:
                     n_discarded_days = 0
 
-                data._days_to_maturity.append(
+                data.days_to_maturity.append(
                     (strategy.expiration - inputs.start_date).days
                     + 1
                     - n_discarded_days
                 )
 
-                data._use_bs.append(strategy.expiration != inputs.target_date)
+                data.use_bs.append(strategy.expiration != inputs.target_date)
             elif isinstance(strategy.expiration, int):
                 if strategy.expiration >= data.days_to_target:
-                    data._days_to_maturity.append(strategy.expiration)
+                    data.days_to_maturity.append(strategy.expiration)
 
-                    data._use_bs.append(strategy.expiration != data.days_to_target)
+                    data.use_bs.append(strategy.expiration != data.days_to_target)
                 else:
                     raise ValueError(
                         "Days remaining to maturity must be greater than or equal to the number of days remaining to the target date!"
@@ -133,20 +127,20 @@ def _init_inputs(inputs: Inputs) -> EngineData:
         elif isinstance(strategy, Stock):
             data.n.append(strategy.n)
             data.action.append(strategy.action)
-            data._previous_position.append(strategy.prev_pos or 0.0)
+            data.previous_position.append(strategy.prev_pos or 0.0)
             data.strike.append(0.0)
             data.premium.append(0.0)
-            data._use_bs.append(False)
-            data._days_to_maturity.append(-1)
+            data.use_bs.append(False)
+            data.days_to_maturity.append(-1)
 
         elif isinstance(strategy, ClosedPosition):
-            data._previous_position.append(strategy.prev_pos)
+            data.previous_position.append(strategy.prev_pos)
             data.strike.append(0.0)
             data.n.append(0)
             data.premium.append(0.0)
             data.action.append("n/a")
-            data._use_bs.append(False)
-            data._days_to_maturity.append(-1)
+            data.use_bs.append(False)
+            data.days_to_maturity.append(-1)
         else:
             raise ValueError("Type must be 'call', 'put', 'stock' or 'closed'!")
 
@@ -156,32 +150,13 @@ def _init_inputs(inputs: Inputs) -> EngineData:
 def _run(data: EngineData) -> EngineData:
     inputs = data.inputs
 
-    time_to_target = data.days_to_target / data._days_in_year
+    time_to_target = data.days_to_target / data.days_in_year
     data.cost = [0.0] * len(data.type)
 
     data.profit = zeros((len(data.type), data.stock_price_array.shape[0]))
     data.strategy_profit = zeros(data.stock_price_array.shape[0])
-    # TODO: To be removed in a next version
-    # if inputs.compute_expectation and data.terminal_stock_prices.shape[0] == 0:
-    #     if inputs.model in ("normal", "black-scholes"):
-    #         terminal_prices_inputs = BlackScholesModelInputs(
-    #             stock_price=inputs.stock_price,
-    #             volatility=inputs.volatility,
-    #             years_to_target_date=time_to_target,
-    #             interest_rate=inputs.interest_rate,
-    #             dividend_yield=inputs.dividend_yield,
-    #         )
-    #     elif inputs.model == "laplace":
-    #         terminal_prices_inputs = LaplaceInputs(
-    #             stock_price=inputs.stock_price,
-    #             volatility=inputs.volatility,
-    #             years_to_target_date=time_to_target,
-    #             mu=inputs.mu,
-    #         )
 
-    #     data.terminal_stock_prices = create_price_samples(terminal_prices_inputs)
-    #
-    if data.terminal_stock_prices.shape[0] > 0:
+    if inputs.model == "array":
         data.profit_mc = zeros((len(data.type), data.terminal_stock_prices.shape[0]))
         data.strategy_profit_mc = zeros(data.terminal_stock_prices.shape[0])
 
@@ -198,10 +173,6 @@ def _run(data: EngineData) -> EngineData:
         if inputs.model == "array":
             data.strategy_profit_mc += data.profit_mc[i]
 
-    data._profit_ranges = get_profit_range(data.stock_price_array, data.strategy_profit)
-
-    pop_inputs: BlackScholesModelInputs | LaplaceInputs | ArrayInputs
-
     if inputs.model in ("normal", "black-scholes"):
         pop_inputs = BlackScholesModelInputs(
             stock_price=inputs.stock_price,
@@ -210,31 +181,39 @@ def _run(data: EngineData) -> EngineData:
             interest_rate=inputs.interest_rate,
             dividend_yield=inputs.dividend_yield,
         )
-    elif inputs.model == "laplace":
-        pop_inputs = LaplaceInputs(
-            stock_price=inputs.stock_price,
-            volatility=inputs.volatility,
-            years_to_target_date=time_to_target,
-            mu=inputs.mu,
-        )
     elif inputs.model == "array":
-        pop_inputs = ArrayInputs(array=data.terminal_stock_prices)
+        pop_inputs = ArrayInputs(array=data.strategy_profit_mc)
     else:
         raise ValueError("Model is not valid!")
 
-    data.profit_probability = get_pop(data._profit_ranges, pop_inputs)
+    pop_out = get_pop(data.stock_price_array, data.strategy_profit, pop_inputs)
 
-    if inputs.profit_target is not None:
-        data._profit_target_range = get_profit_range(
-            data.stock_price_array, data.strategy_profit, inputs.profit_target
-        )
-        data.profit_target_probability = get_pop(data._profit_target_range, pop_inputs)
+    data.profit_probability = pop_out.probability_of_reaching_target
+    data.expected_profit = pop_out.expected_return_above_target
+    data.expected_loss = pop_out.expected_return_below_target
+    data.profit_ranges = pop_out.reaching_target_range
 
-    if inputs.loss_limit is not None:
-        data._loss_limit_ranges = get_profit_range(
-            data.stock_price_array, data.strategy_profit, inputs.loss_limit + 0.01
+    if inputs.profit_target is not None and inputs.profit_target > 0.01:
+        pop_out_prof_targ = get_pop(
+            data.stock_price_array,
+            data.strategy_profit,
+            pop_inputs,
+            inputs.profit_target,
         )
-        data.loss_limit_probability = 1.0 - get_pop(data._loss_limit_ranges, pop_inputs)
+        data.profit_target_probability = (
+            pop_out_prof_targ.probability_of_reaching_target
+        )
+        data.profit_target_ranges = pop_out_prof_targ.reaching_target_range
+
+    if inputs.loss_limit is not None and inputs.loss_limit < 0.0:
+        pop_out_loss_lim = get_pop(
+            data.stock_price_array,
+            data.strategy_profit,
+            pop_inputs,
+            inputs.loss_limit + 0.01,
+        )
+        data.loss_limit_probability = pop_out_loss_lim.probability_of_missing_target
+        data.loss_limit_ranges = pop_out_loss_lim.missing_target_range
 
     return data
 
@@ -244,7 +223,7 @@ def _run_option_calcs(data: EngineData, i: int) -> EngineData:
     action: Action = data.action[i]  # type: ignore
     type: OptionType = data.type[i]  # type: ignore
 
-    if data._previous_position[i] < 0.0:
+    if data.previous_position[i] < 0.0:
         # Previous position is closed
         data.implied_volatility.append(0.0)
         data.itm_probability.append(0.0)
@@ -254,7 +233,7 @@ def _run_option_calcs(data: EngineData, i: int) -> EngineData:
         data.theta.append(0.0)
         data.rho.append(0.0)
 
-        cost = (data.premium[i] + data._previous_position[i]) * data.n[i]
+        cost = (data.premium[i] + data.previous_position[i]) * data.n[i]
 
         if data.action[i] == "buy":
             cost *= -1.0
@@ -267,7 +246,7 @@ def _run_option_calcs(data: EngineData, i: int) -> EngineData:
 
         return data
 
-    time_to_maturity = data._days_to_maturity[i] / data._days_in_year
+    time_to_maturity = data.days_to_maturity[i] / data.days_in_year
     bs = get_bs_info(
         inputs.stock_price,
         data.strike[i],
@@ -297,23 +276,23 @@ def _run_option_calcs(data: EngineData, i: int) -> EngineData:
     if type == "call":
         data.itm_probability.append(bs.call_itm_prob)
         data.delta.append(bs.call_delta * negative_multiplier)
-        data.theta.append(bs.call_theta / data._days_in_year * negative_multiplier)
+        data.theta.append(bs.call_theta / data.days_in_year * negative_multiplier)
         data.rho.append(bs.call_rho * negative_multiplier)
     else:
         data.itm_probability.append(bs.put_itm_prob)
         data.delta.append(bs.put_delta * negative_multiplier)
-        data.theta.append(bs.put_theta / data._days_in_year * negative_multiplier)
+        data.theta.append(bs.put_theta / data.days_in_year * negative_multiplier)
         data.rho.append(bs.put_rho * negative_multiplier)
 
-    if data._previous_position[i] > 0.0:  # Premium of the open position
-        opt_value = data._previous_position[i]
+    if data.previous_position[i] > 0.0:  # Premium of the open position
+        opt_value = data.previous_position[i]
     else:  # Current premium
         opt_value = data.premium[i]
 
-    if data._use_bs[i]:
+    if data.use_bs[i]:
         target_to_maturity = (
-            data._days_to_maturity[i] - data.days_to_target
-        ) / data._days_in_year  # To consider the expiration date as a trading day
+            data.days_to_maturity[i] - data.days_to_target
+        ) / data.days_in_year  # To consider the expiration date as a trading day
 
         data.profit[i], data.cost[i] = get_pl_profile_bs(
             type,
@@ -384,8 +363,8 @@ def _run_stock_calcs(data: EngineData, i: int) -> EngineData:
     data.rho.append(0.0)
     data.theta.append(0.0)
 
-    if data._previous_position[i] < 0.0:  # Previous position is closed
-        costtmp = (inputs.stock_price + data._previous_position[i]) * data.n[i]
+    if data.previous_position[i] < 0.0:  # Previous position is closed
+        costtmp = (inputs.stock_price + data.previous_position[i]) * data.n[i]
 
         if data.action[i] == "buy":
             costtmp *= -1.0
@@ -398,8 +377,8 @@ def _run_stock_calcs(data: EngineData, i: int) -> EngineData:
 
         return data
 
-    if data._previous_position[i] > 0.0:  # Stock price at previous position
-        stockpos = data._previous_position[i]
+    if data.previous_position[i] > 0.0:  # Stock price at previous position
+        stockpos = data.previous_position[i]
     else:  # Spot price of the stock at start date
         stockpos = inputs.stock_price
 
@@ -434,63 +413,36 @@ def _run_closed_position_calcs(data: EngineData, i: int) -> EngineData:
     data.rho.append(0.0)
     data.theta.append(0.0)
 
-    data.cost[i] = data._previous_position[i]
-    data.profit[i] += data._previous_position[i]
+    data.cost[i] = data.previous_position[i]
+    data.profit[i] += data.previous_position[i]
 
     if inputs.model == "array":
-        data.profit_mc[i] += data._previous_position[i]
+        data.profit_mc[i] += data.previous_position[i]
 
     return data
 
 
 def _generate_outputs(data: EngineData) -> Outputs:
-    inputs = data.inputs
-    optional_outputs: dict[str, Any] = {}
-
-    if inputs.profit_target is not None:
-        optional_outputs["probability_of_profit_target"] = (
-            data.profit_target_probability
-        )
-        optional_outputs["profit_target_ranges"] = data._profit_target_range
-
-    if inputs.loss_limit is not None:
-        optional_outputs["probability_of_loss_limit"] = data.loss_limit_probability
-
-    if inputs.model == "array" and data.terminal_stock_prices.shape[0] > 0:
-        profit = data.strategy_profit_mc[data.strategy_profit_mc >= 0.01]
-        loss = data.strategy_profit_mc[data.strategy_profit_mc < 0.0]
-        optional_outputs["average_profit_from_mc"] = 0.0
-        optional_outputs["average_loss_from_mc"] = (
-            loss.mean() if loss.shape[0] > 0 else 0.0
-        )
-
-        if profit.shape[0] > 0:
-            optional_outputs["average_profit_from_mc"] = profit.mean()
-
-        if loss.shape[0] > 0:
-            optional_outputs["average_loss_from_mc"] = loss.mean()
-
-        optional_outputs["probability_of_profit_from_mc"] = (
-            data.strategy_profit_mc >= 0.01
-        ).sum() / data.strategy_profit_mc.shape[0]
-
-    return Outputs.model_validate(
-        optional_outputs
-        | {
-            "inputs": inputs,
-            "data": data,
-            "probability_of_profit": data.profit_probability,
-            "strategy_cost": sum(data.cost),
-            "per_leg_cost": data.cost,
-            "profit_ranges": data._profit_ranges,
-            "minimum_return_in_the_domain": data.strategy_profit.min(),
-            "maximum_return_in_the_domain": data.strategy_profit.max(),
-            "implied_volatility": data.implied_volatility,
-            "in_the_money_probability": data.itm_probability,
-            "delta": data.delta,
-            "gamma": data.gamma,
-            "theta": data.theta,
-            "vega": data.vega,
-            "rho": data.rho,
-        }
+    return Outputs(
+        inputs=data.inputs,
+        data=data,
+        probability_of_profit=data.profit_probability,
+        expected_profit=data.expected_profit,
+        expected_loss=data.expected_loss,
+        strategy_cost=sum(data.cost),
+        per_leg_cost=data.cost,
+        profit_ranges=data.profit_ranges,
+        minimum_return_in_the_domain=data.strategy_profit.min(),
+        maximum_return_in_the_domain=data.strategy_profit.max(),
+        implied_volatility=data.implied_volatility,
+        in_the_money_probability=data.itm_probability,
+        delta=data.delta,
+        gamma=data.gamma,
+        theta=data.theta,
+        vega=data.vega,
+        rho=data.rho,
+        probability_of_profit_target=data.profit_target_probability,
+        probability_of_loss_limit=data.loss_limit_probability,
+        profit_target_ranges=data.profit_target_ranges,
+        loss_limit_ranges=data.loss_limit_ranges,
     )